Native Runner for CPP Codes
This is a modified version of the testAnaSingleThread.py file that is written in CPP, and if inserted into the same repository as the kuznetsov-radio CPP codes, allows for the running of the codes from a CPP compiler, such as Visual Studio. It cannot produce graphs, although, the data it returns from the codes is computed quickly and is accurate. This script can be modified to output its results to a file which can later be read by another program, and plotted by something like python's matplotlib.